Change namespace Lightna\Magento -> Lightna\Magento\Backend. Rename rules -> rule Fix phpdoc usage. Fix EntityRoute storage. ./cli update -> ./cli schema.update. fix error for ./cli index.entity.show route -s1 "": Specify entity ID. Make better error message when bin/magento fails and there is not lightna build. Fix fail when no static block managed. Project from scratch: fix error " table lightna_indexer_changelog doesn't exist".

+++ code/lightna/lightna-engine/App/Query/Index/Changelog.php
@@ -9,7 +9,7 @@ use Laminas\Db\Sql\Select;
 use Laminas\Db\Sql\Update;
 use Lightna\Engine\App\ObjectA;
 use Lightna\Engine\App\Project\Database;
-use Lightna\Engine\App\Update\Schema\Index\Changelog as Schema;
+use Lightna\Engine\App\Schema\Index\Changelog as Schema;
 
 class Changelog extends ObjectA
 {
@@ -119,6 +119,10 @@ class Changelog extends ObjectA
 
     public function reset(): void
     {
+        if (!$this->db->isUsed()) {
+            return;
+        }
+
         $this->db->query('truncate table ' . $this->db->quoteIdentifier(Schema::TABLE_NAME));
     }
